fix conflicts

+++ src/store/modules/authentication.js
@@ -22,27 +22,28 @@ export const mutations = {
 export const actions = {
     login({ commit }, userData) {
         return StreamsServices.postLogin(userData).then(response => {
-            response.data = {success:true,token:"testtoken"};//TODO
-            commit("SET_ALL_UserData",{
-                success:response.data.success,
-                name:userData.username,
-                token:response.data.token
+            commit("SET_ALL_UserData", {
+                success: response.data.success,
+                name: userData.username,
+                token: response.data.userdata.token
             });
-            if(response.data.success)
-                CookieService.setToken(response.data.token)
+            if (response.data.success)
+                CookieService.setToken(response.data.userdata.token)
             return response.data.success;
-        });
+        }).catch(() => {return false});
     },
-    authenticate({ commit, dispatch },token){
+    authenticate({ commit, dispatch }){
+        let token = CookieService.getToken();
+        if (!token) return false;
         return StreamsServices.postToken(token).then(response => {
-            response.data = {success:true,username:"testuser"}   //TODO
-            if(response.data.success){
-                commit("SET_ALL_UserData",{token:token,success:true,name:response.username}) 
-            }else{
-                commit("SET_ALL_UserData",{token:null,loggedIn:false,name:null})
+            if (response.data.success) {
+                commit("SET_ALL_UserData", { token: token, success: true, name: response.data.userdata.username })
+                return true;
+            } else {
+                commit("SET_ALL_UserData", { token: null, loggedIn: false, name: null })
                 CookieService.removeToken();
+                return false;
             }
-            return response.data.success;
         }).catch(error => {
             const notification = {
                 type: "error",
